requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.9
django-comments-xtd==2.10.0
django-contrib-comments==2.2.0
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
docutils==0.21.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
